+++ src/smoke.css
@@ -81,7 +81,7 @@
   position: absolute;
   width: 2px;
   height: 10px;
-  background: rgba(255, 255, 255, 0.7);
+  background: rgba(255, 255, 255, 0.663);
   animation: rainAnimation 0.5s linear infinite;
 }
